{{cookiecutter.project_slug}}/requirements/base.txt
certifi==2017.11.5
chardet==3.0.4
diff-match-patch==20121119
Django==1.11.8
django-bower==5.2.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==1.9.8
django-formtools==2.1
django-import-export==0.6.1
django-maintenancemode-2==1.1.9
django-mathjax==0.0.8
django-multiselectfield==0.1.8
django-recaptcha==1.3.1
django-reversion==2.0.12
django-user-agents==0.3.2
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.6
jdcal==1.3
odfpy==1.3.6
olefile==0.44
openpyxl==2.4.9
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
PyYAML==3.12
requests==2.18.4
six==1.11.0
sparkpost==1.3.5
sqlparse==0.2.4
tablib==0.12.1
typing==3.6.2
ua-parser==0.7.3
unicodecsv==0.14.1
urllib3==1.22
user-agents==1.1.0
Werkzeug==0.13
git+git://github.com/engineervix/xadmin.git
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
